requirements.txt
asgiref==3.8.1
attrs==25.3.0
colorama==0.4.6
Django==5.2.1
django-debug-toolbar==5.2.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Markdown==3.8
packaging==25.0
phonenumbers==9.0.5
pillow==11.2.1
pluggy==1.6.0
psycopg==3.2.8
psycopg-binary==3.2.8
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
